pyproject.toml
[tool.poetry]
name = "hebcal"
version = "0.1.0"
description = ""
authors = ["meirdev <maornet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyluach = "^2.2.0"
python-dateutil = "^2.8.2"
astral = "^3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
